+++ doc/whats-new.md
@@ -137,6 +137,7 @@ Furthermore, you can now bind any keyboard keys (not just 0-9/A-Z).  See [HotKey
  - when you get a reminder to consider the magical sword, tracker now briefly highlights its location on the map
  - when you get the 8th triforce piece, reminder that dungeon 9 is unlocked now briefly highlights the location of dungeon 9
  - there's now reminders for Door Repair Count on the Options Menu
+ - you can now choose to hide certain overworld icons you don't want to see (Options Menu\Overworld\More settings)
  - when you reset the timer, you immediately get put into the 'place the start spot location icon' popup to mark your start screen
  - you can now middle-click-drag across multiple dungeon rooms to mark them as 'off the map' spots
  - Show/Run Custom button now can 'RUN' URLs, which get launched in your default browser